refactor: move device-deletion logic to SongRemovalStateHolder
Relocates the implementation of single and batch file deletions from `PlayerViewModel` to `SongRemovalStateHolder` to improve separation of concerns and reduce ViewModel complexity.

- Created `SongRemovalCallbacks` to provide `SongRemovalStateHolder` with access to ViewModel-owned operations (toasts, queue management, and library removal) without a direct circular dependency.
- Moved `MediaStore` delete-permission request logic and pending deletion state management.
- Migrated the implementation of `deleteSelectedFromDevice`, `deleteFromDevice`, and `onDeletePermissionResult`, including confirmation dialog logic.
- Updated `PlayerViewModel` to delegate deletion actions to the state holder and re-expose the permission request flow to the UI.
